requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
jinja2
